income-platform/src/admin-panel/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
jinja2==3.1.6
httpx==0.28.1
PyJWT==2.10.1
SQLAlchemy==2.0.40
psycopg2-binary==2.9.10
python-multipart==0.0.20
